+++ kaiba/models/regex.py
@@ -8,5 +8,5 @@ from kaiba.models.base import KaibaBaseModel
 class Regex(KaibaBaseModel):
     """Use regular expression to find your data."""
 
-    search: Pattern
+    expression: Pattern
     group: Union[StrictInt, List[StrictInt]] = 0