+++ tests/validation/test_cast.py
@@ -1,22 +1,22 @@
 import pytest
 from pydantic import ValidationError
 
-from kaiba.pydantic_schema import Casting, CastingEnum
+from kaiba.models.cast import Cast, CastToOptions
 
 
 def test_validates():  # noqa: WPS218
     """Test that dict is marshalled to pydantic object."""
-    test = Casting(
+    test = Cast(
         to='integer',
     )
-    assert test.to == CastingEnum.INTEGER
+    assert test.to == CastToOptions.INTEGER
     assert test.original_format is None
 
 
 def test_invalid():
     """Test that we get validation error with correct message."""
     with pytest.raises(ValidationError) as ve:
-        Casting(to='test')
+        Cast(to='test')
 
     errors = ve.value.errors()[0]  # noqa: WPS441
     assert errors['loc'] == ('to',)